+++ .gitignore
@@ -1,7 +1,9 @@
-# See https://help.github.com/ignore-files/ for more about ignoring files.
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
 /node_modules
+/.pnp
+.pnp.js
 
 # testing
 /coverage
@@ -20,11 +22,6 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-# workflow
-.netlify
+# Environment
 .vscode
-
-# production
-build
-.next
-package-lock.json
+.netlify
