+++ package.json
@@ -1,25 +1,31 @@
 {
-  "name": "site",
+  "name": "personal",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "next": "^8.1.0",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
-    "styled-components": "^4.2.0"
-  },
-  "devDependencies": {
-    "babel-plugin-styled-components": "^1.8.0"
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-scripts": "3.0.1"
   },
   "scripts": {
-    "dev": "next",
-    "build": "next build",
-    "start": "next start"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
