--- pages/_app.js
@@ -1,32 +1,0 @@
-import App, { Container } from 'next/app';
-import React from 'react';
-import { ThemeProvider } from 'styled-components';
-
-const theme = {
-  colors: {
-    primary: '#0070f3'
-  }
-};
-
-export default class MyApp extends App {
-  static async getInitialProps({ Component, ctx }) {
-    let pageProps = {};
-
-    if (Component.getInitialProps) {
-      pageProps = await Component.getInitialProps(ctx);
-    }
-
-    return { pageProps };
-  }
-
-  render() {
-    const { Component, pageProps } = this.props;
-    return (
-      <Container>
-        <ThemeProvider theme={theme}>
-          <Component {...pageProps} />
-        </ThemeProvider>
-      </Container>
-    );
-  }
-}
