+++ src/components/projects/ProjectDetails.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useRef } from 'react';
 import {
   Container,
   Picture,
@@ -7,10 +7,12 @@ import {
 } from '../styled-components/projects/project-details';
 
 function ProjectDetails(props) {
+  const imageEl = useRef(null);
   const { image, name, description, tecnologies } = props;
   return (
     <Container>
-      <Img src={image} alt={name} />
+      <Img ref={imageEl} src={image} alt={name} />
+      <div>Height: {height}</div>
       {/* <Content>
         <div>{name}</div>
       </Content> */}
