+++ src/components/styled-components/footer/footer-icon-details.js
