+++ src/components/styled-components/introduction/introduction-details.js
