+++ src/components/styled-components/projects/project-details.js
